backend/requirements.txt
Django>=5.0
 djangorestframework>=3.14
 djangorestframework-simplejwt>=5.3
 django-cors-headers>=4.3
 python-dotenv>=1.0
